feat: polish game topnav controls
