+++ app/src/renderer/windows/game/style.css
@@ -53,7 +53,7 @@
 .game-topnav__select-trigger {
   align-items: center;
   background: transparent;
-  border: 0;
+  border: 1px solid transparent;
   border-radius: var(--radius-sm);
   color: var(--color-muted-foreground);
   cursor: default;
@@ -65,23 +65,47 @@
   height: 1.5rem;
   justify-content: center;
   line-height: 1;
+  position: relative;
   min-width: 0;
   padding: 0 0.625rem;
+  transition:
+    background var(--duration-fast) ease,
+    border-color var(--duration-fast) ease,
+    box-shadow var(--duration-fast) ease,
+    color var(--duration-fast) ease,
+    opacity var(--duration-fast) ease;
   white-space: nowrap;
 }
 
 .game-topnav__trigger:hover,
-.game-topnav__trigger[data-expanded],
 .game-topnav__button:not(:disabled):hover,
 .game-topnav__select-trigger:not(:disabled):hover {
-  background: var(--color-accent);
+  background: color-mix(in srgb, var(--color-accent) 82%, var(--color-background));
   color: var(--color-foreground);
 }
 
-.game-topnav__left .game-topnav__trigger:focus,
-.game-topnav__left .game-topnav__button:focus {
-  box-shadow: none;
+.game-topnav__trigger[data-expanded],
+.game-topnav__select-trigger[data-expanded] {
+  background: color-mix(in srgb, var(--color-primary) 16%, var(--color-accent));
+  border-color: color-mix(in srgb, var(--color-primary) 48%, transparent);
+  box-shadow:
+    inset 0 -2px 0 var(--color-primary),
+    0 0 0 1px color-mix(in srgb, var(--color-primary) 18%, transparent);
+  color: var(--color-foreground);
+}
+
+.game-topnav__trigger:focus-visible,
+.game-topnav__button:focus-visible,
+.game-topnav__select-trigger:focus-visible {
+  background: color-mix(in srgb, var(--color-ring) 14%, var(--color-accent));
+  border-color: color-mix(in srgb, var(--color-ring) 72%, transparent);
+  box-shadow:
+    0 0 0 2px var(--color-background),
+    0 0 0 4px color-mix(in srgb, var(--color-ring) 88%, transparent),
+    inset 0 -2px 0 var(--color-ring);
+  color: var(--color-foreground);
   outline: none;
+  z-index: 1;
 }
 
 .game-topnav__trigger--success,
@@ -94,14 +118,32 @@
 }
 
 .game-topnav__select-trigger {
-  border: 1px solid var(--color-border);
+  border: 1px solid color-mix(in srgb, var(--color-border) 92%, transparent);
+  box-shadow: inset 0 1px 0 color-mix(in srgb, var(--color-foreground) 8%, transparent);
   font-size: var(--text-xs);
-  justify-content: flex-start;
-  width: 4rem;
+  justify-content: space-between;
+  max-width: 7rem;
+  min-width: 5.5rem;
+  overflow: hidden;
+  padding-right: 0.375rem;
+  text-overflow: ellipsis;
+  width: 5.5rem;
+}
+
+.game-topnav__select-trigger::after {
+  border-left: 0.25rem solid transparent;
+  border-right: 0.25rem solid transparent;
+  border-top: 0.3125rem solid currentColor;
+  content: "";
+  flex: 0 0 auto;
+  margin-left: 0.375rem;
+  opacity: 0.55;
 }
 
 .game-topnav__select-trigger--cell {
-  width: 5rem;
+  max-width: 9rem;
+  min-width: 7rem;
+  width: 7rem;
 }
 
 .game-topnav__trigger:disabled,
@@ -128,7 +170,7 @@
 }
 
 .game-menu--mega {
-  min-width: min(34rem, calc(100vw - 1rem));
+  min-width: min(30rem, calc(100vw - 1rem));
 }
 
 .game-menu--scripts,
@@ -144,10 +186,18 @@
   min-width: 7rem;
 }
 
+.game-menu--pads {
+  min-width: 8rem;
+}
+
+.game-menu--cells {
+  min-width: 10rem;
+}
+
 .game-menu__mega-grid {
   display: grid;
-  gap: 0.25rem;
-  grid-template-columns: repeat(3, minmax(0, 1fr));
+  gap: 0.375rem;
+  grid-template-columns: repeat(2, minmax(12rem, 1fr));
 }
 
 .game-menu__group {
@@ -156,6 +206,23 @@
 
 .game-menu__item {
   min-width: 0;
+}
+
+.game-menu__pad-option {
+  align-items: baseline;
+}
+
+.game-menu__pad-name {
+  flex: 1 1 auto;
+  font-weight: 600;
+  min-width: 0;
+  overflow: hidden;
+  text-overflow: ellipsis;
+  white-space: nowrap;
+}
+
+.game-menu__pad-option--valid .game-menu__pad-name {
+  color: rgb(var(--success-foreground));
 }
 
 .game-menu__item-label {
