+++ app/src/shared/windows.ts
@@ -80,14 +80,14 @@ export const gameWindowGroups: readonly WindowGroup[] = [
         },
       },
       {
-        id: WindowIds.FastTravels,
-        label: "Fast travels",
-        description: "Open fast-travel helpers for the active game window.",
+        id: WindowIds.Follower,
+        label: "Follower",
+        description: "Coordinate follower behavior for the active game.",
         scope: "game-child",
         closeBehavior: "hide",
         dimensions: {
-          width: 649,
-          height: 527,
+          width: 648,
+          height: 496,
         },
       },
       {
@@ -102,14 +102,14 @@ export const gameWindowGroups: readonly WindowGroup[] = [
         },
       },
       {
-        id: WindowIds.Follower,
-        label: "Follower",
-        description: "Coordinate follower behavior for the active game.",
+        id: WindowIds.FastTravels,
+        label: "Fast travels",
+        description: "Open fast-travel helpers for the active game window.",
         scope: "game-child",
         closeBehavior: "hide",
         dimensions: {
-          width: 648,
-          height: 496,
+          width: 649,
+          height: 527,
         },
       },
     ],
